+++ miracle/worker/app.py
@@ -9,6 +9,7 @@ from kombu import Queue
 from miracle.bloom import create_bloom_domain
 from miracle.bucket import create_bucket
 from miracle.cache import create_cache
+from miracle.crypto import create_crypto
 from miracle.db import create_db
 from miracle.config import REDIS_URI
 from miracle.log import (
@@ -36,7 +37,7 @@ def configure_celery(celery_app):
 
 def init_worker(celery_app,
                 _bloom_domain=None, _bucket=None, _cache=None,
-                _db=None, _raven=None, _stats=None):
+                _crypto=None, _db=None, _raven=None, _stats=None):
     configure_logging()
     raven = create_raven(transport='threaded', _raven=_raven)
 
@@ -44,6 +45,7 @@ def init_worker(celery_app,
         celery_app.bloom_domain = create_bloom_domain(_bloom=_bloom_domain)
         celery_app.bucket = create_bucket(_bucket=_bucket)
         celery_app.cache = create_cache(_cache=_cache)
+        celery_app.crypto = create_crypto(_crypto=_crypto)
         celery_app.db = create_db(_db=_db)
         celery_app.raven = raven
         celery_app.stats = create_stats(_stats=_stats)
@@ -61,6 +63,7 @@ def shutdown_worker(celery_app):
     del celery_app.bucket
     celery_app.cache.close()
     del celery_app.cache
+    del celery_app.crypto
     celery_app.db.close()
     del celery_app.db
     del celery_app.raven